Aula04_Project_Agent_Fullstack/requirements.txt
langchain
langchain-core
langchain-openai
langchain-mongodb
openai
fastapi
uvicorn
python-dotenv
pymongo
datasets
websockets
uv
